changing directions and adding mafia route befor pages
