+++ pages/mafia/conductor/event/create.js
@@ -7,14 +7,14 @@ import 'react-toastify/dist/ReactToastify.css';
 import { useRouter } from "next/router";
 import Link from "next/link";
 import { getCookie } from "cookies-next";
-import checkToken from "../../../utils/checkToken";
-import Nav from "../../../components/nav";
-import Header from "../../../components/header";
-import createEvent from "../../../utils/createEvent";
+import checkToken from "../../../../utils/checkToken";
+import Nav from "../../../../components/nav";
+import Header from "../../../../components/header";
+import createEvent from "../../../../utils/createEvent";
 import { useSelector } from "react-redux";
-import Circular from "../../../components/Circular";
+import Circular from "../../../../components/Circular";
 import DateObject from "react-date-object";
-import DateInput from "../../../components/DateInput";
+import DateInput from "../../../../components/DateInput";
 
 const Create = props => {
     const Router = useRouter()
@@ -30,7 +30,7 @@ const Create = props => {
         else {
             toast.error("وارد حساب کاربری خود شوید")
             setTimeout(() => {
-                Router.push('/')
+                Router.push('/mafia/login')
             }, 2000)
         }
     }, [globalUser])
@@ -177,7 +177,7 @@ const Create = props => {
 
         if (status === 201 || status === 302) {
             toast.success('ایونت با موفقیت ایجاد شد')
-            setTimeout(() => Router.push('/conductor/event/'), 2000)
+            setTimeout(() => Router.push('/mafia/conductor/event/'), 2000)
         } else {
             toast.error('خطایی در هنگام ایجاد ایونت بوجود آمده')
             button.removeAttribute('disabled')
@@ -203,7 +203,7 @@ const Create = props => {
 
                         <div className="page-title">
                             <h2>افزودن ایونت</h2>
-                            <Link href={'/conductor/event'}>
+                            <Link href={'/mafia/conductor/event'}>
                                 <a>
                                     بازگشت
                                 </a>
