+++ pages/mafia/events/[id]/[time]/[phase].js
